feat: Added area information component and ngrx support.
The main container component initializes the store by dispatching actions.  Since we need to pull the current area information out of the store at initialization, added mergeMap to the inital AREA_LIST effect so that area information is added later in the stream, after the service responds with data. Since the container component dispatches a redundant AreaInformationUpdate action at initialization, also added an effect that checks for empty store and switches to failure action if no data is available.  This is under review.

Issues: AC-917