feat: Component unit tests updated to work with angular 4.
The tests appeared to work with webpack afer ng serve, but continued to fail with ng test.  The issue appeared to be with the use of mock classes in the testbed providers.  Swapping these out in favor of anonymous classes fixed the testbed problem. Tests probably should be reviewed later once depdencies have settled down.

Issues: AC-919