feat: Added CAS authentication and authentication check.
The current state of Angular4 and third-party middleware makes it difficult to check the session cookie after a CAS login.  For now, I added an endpoint that can be used by the ItemLinksComponent to determine authenticatin status.  This is roundabout and lame.  But will work fine until a better solution comes along.  Changed the Pastport CAS middleware to a strategy that validates and removes the CAS token parameter from the url.

Issues: AC-919