fix: Corrected problem with displaying items to which no subjects are attached.
Without actually dispatching a related collections request my item data subscription was not binding properly to the template.  Reverted back to letting the framework bind the observable using async. Also created a separate reference for a subscriber to the item store.  This subscription handles the related items request and is unsubscribed in onDestroy.  For clarity, also created a separate actions module for related collections.

Issues: AC-1023