fix: Corrects problem in the select options component.
Previously the md-select directive did not respond to initial click events but did respond
once the user had clicked somewhere on the page. The solution was to relocate the service request
for options data from the API and provide the Observable to the *ngFor loop using the async pipe.

Issue: AC-1047