bugfix: It's now possible to remove type filter when in mobile view.
The type filter is not available in mobile, but it is possible to deep link into a list that is filtered by type.  The function that checks for the presenece of an id in the list of type objects failed.  Added a test for an undefined object array.

Issues: AC-1150